requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.6
bcrypt==3.2.0
djangorestframework_simplejwt=5.2.2
pyjwt=2.8.0
